Updating EMsoftCpreprocessEBSDPatterns callback to take an int32_t iPar pointer.
The corresponding Fortran method declares the iPar array as a 32-bit integer array, and there are issues if we send in a size_t array instead.
This commit fixes the callback so that it is consistent with its corresponding Fortran method.

+++ Source/EMsoftWrapperLib/DictionaryIndexing/EMsoftDIwrappers.h
@@ -43,7 +43,7 @@ typedef void (*ProgCallBackTypeError)(size_t, int);
 * @param cancel boolean to trigger cancellation of computation
 */
 void EMsoftCpreprocessEBSDPatterns
-    (size_t* ipar, float* fpar, char* spar, float* mask, 
+    (int32_t* ipar, float* fpar, char* spar, float* mask,
      float* exptIQ, float* ADPmap, ProgCallBackTypeDI2 callback, 
      size_t object, bool* cancel);
